move country.{status,config,desk_study} to country.props
