+++ meta/assessment/index.ts
@@ -4,13 +4,11 @@ export { ColType } from './col'
 export { Cols } from './cols'
 export { RowType } from './row'
 export { Unit } from './unit'
-export { AssessmentStatus } from './countryStatus'
 export { ODPNationalClassFactory, ODPs, ODPYears } from './originalDataPoint'
 
 export type { ODPNationalClass } from './originalDataPoint'
 export type { ActivityLog } from './activityLog'
 export type { Assessment } from './assessment'
-export type { CountryStatus } from './countryStatus'
 export type { Col, ColProps } from './col'
 export type { CycledPropsObject, Cycle } from './cycle'
 export type { Descriptions, Description } from './description'
